// FASTraxCoreWeb/src/helpers/asset/enum.ts
export enum AssetCategoryEnum {
    GeneralAsset = 1,
    Vehicle = 2,
    Component = 3    
}

export enum TransactionHistoryTypeEnum {
    Disposition = 1,
    Estimate = 2,
    Inspection = 3,    
    WorkOrder = 4,
}

export enum DispositionStatusEnum {
    Open = 1,
    InProgress = 2,
    Approved = 3,
    Rejected = 4
}